requirements.txt
fastapi
uvicorn
websockets
camoufox[geoip]
playwright
python-dotenv
supabase
groq
